requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-cors==0.8.1
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
astor==0.8.1
async-timeout==5.0.1
attrs==25.3.0
av==15.1.0
blake3==1.0.5
cachetools==5.5.2
cbor2==5.7.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
codetiming==1.4.0
colorful==0.5.7
compressed-tensors==0.10.2
cupy-cuda12x==13.6.0
datasets==4.0.0
depyf==0.19.0
diffusers @ git+https://github.com/huggingface/diffusers@a519272d97f011332588e1aaa73d32952d80f3f5
dill==0.3.8
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
dnspython==2.7.0
einops==0.8.1
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.116.1
fastapi-cli==0.0.10
fastapi-cloud-cli==0.1.5
fastrlock==0.8.3
filelock==3.19.1
flash_attn==2.8.0.post2
frozenlist==1.7.0
fsspec==2025.3.0
gguf==0.17.1
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.74.0
h11==0.16.0
hf-xet==1.1.9
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.7.0
interegular==0.3.3
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
lark==1.2.2
liger_kernel==0.6.2
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.10.12
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mathruler==0.1.0
mdurl==0.1.2
mistral_common==1.8.4
mpmath==1.3.0
msgpack==1.1.1
msgspec==0.19.0
multidict==6.6.4
multiprocess==0.70.16
networkx==3.4.2
ninja==1.13.0
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
openai==1.106.1
openai-harmony==0.0.4
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python-headless==4.12.0.88
opentelemetry-api==1.36.0
opentelemetry-exporter-prometheus==0.57b0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.3
outlines_core==0.2.10
packaging==25.0
pandas==2.3.2
partial-json-parser==0.2.1.1.post6
peft==0.17.1
pillow==11.3.0
platformdirs==4.4.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.1
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.0
psutil==7.0.0
py-cpuinfo==9.0.0
py-spy==0.4.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycountry==24.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic_core==2.33.2
Pygments==2.19.2
pylatexenc==2.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
pyvers==0.1.0
PyYAML==6.0.2
pyzmq==27.0.2
qwen-vl-utils==0.0.11
ray==2.49.1
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
scipy==1.15.3
sentencepiece==0.2.1
sentry-sdk==2.37.0
setproctitle==1.3.7
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
starlette==0.47.3
sympy==1.14.0
tensordict==0.9.1
tiktoken==0.11.0
tokenizers==0.22.0
torch==2.7.1
torchaudio==2.7.1
torchdata==0.11.0
torchvision==0.22.1
tqdm==4.67.1
transformers==4.56.1
triton==3.3.1
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
-e git+https://github.com/hiyouga/EasyR1.git@d3ebe5b8ec9e9c3317a0b8b0e8a48ce4283f6131#egg=verl
virtualenv==20.34.0
vllm==0.10.1.1
wandb==0.21.3
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.3
xformers==0.0.31
xgrammar==0.1.21
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
